# src/main/webapp/openapi.yaml
openapi: 3.0.3
info:
  title: Festival Tracker - OpenAPI 3.0
  description: |-
    This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.

    _If you're looking for the Swagger 2.0/OAS 2.0 version of Petstore, then click [here](https://editor.swagger.io/?url=https://petstore.swagger.io/v2/swagger.yaml). Alternatively, you can load via the `Edit > Load Petstore OAS 2.0` menu option!_
    
    Some useful links:
    - [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)
    - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: http://localhost:8080/EntJavaTeamProject_war/services/
...
paths:
...
  /festivals/types/{typeId}:
    get:
      tags:
        - festival
      summary: Finds Festivals by type
      description: Finds a list of festivals by type
      operationId: findFestivalsByType
      parameters:
        - name: typeId
          in: path
          description: ID of type to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Festival'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Festival'
        '400':
          description: Invalid ID supplied
        '404':
          description: No festivals found
...
  /types/{typeId}:
    get:
      tags:
        - type
      summary: Find type  by ID
      description: find a type by the type id
      operationId: getTypeById
      parameters:
        - name: typeId
          in: path
          description: ID of type that needs to be fetched
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Type'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Type'
        '400':
          description: Invalid ID supplied
        '404':
          description: Region not found
components:
  schemas:
    Type:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        type:
          type: string
          example: 'exampleType'
      xml:
        name: type
...
    Festival:
      required:
        - name
        - typeId
        - regionId
        - date
        - startDate
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
          example: 'exampleFestival'
        typeId:
          type: integer
          format: int64
          example: 4
        regionId:
          type: integer
          format: int64
          example: 7
        startDate: 
          type: string
          format: date  
          example: '2014-11-11'
    # ApiResponse:
    #   type: object
    #   properties:
    #     code:
    #       type: integer
    #       format: int32
    #     type:
    #       type: string
    #     message:
    #       type: string
    #   xml:
    #     name: '##default'
...